split into main.dart and home_page.dart
